% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmi.R
\name{mcmi}
\alias{mcmi}
\title{Monte Carlo Method for Multivariable Integration}
\usage{
mcmi(bounds, integrand, constrains, n)
}
\arguments{
\item{bounds:}{necessary, the bounds of x, y, z, as a vector of 2, 4 or 6 dimensions}

\item{integrand:}{optional, the integrand, input as character, must contains x, y, or z,   default: 1}

\item{constrains:}{optional, inequalities that x, y, z must satisfy, input as character, must contains x, y, or z, default: 1}

\item{n:}{number of random points, default 1,000,000,}
}
\description{
Calculate multivariable integration using Monte Carlo method, only support up to 3D currently, accuracy 0.005
}
\examples{
mcmi(c(-1,1, -1,1), 1, "x^2+y^2<1")
(-1<x<1, -1<y<1)

mcmi(c(0,1, 1,2, -1,3), 1, c("exp(x)<y", "sin(z)*y>=0")
(0<x<1, 1<y<2, -1<3)

mcmi(c(0,1), "exp(x)")
(0<x<1)

}
