% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_area}
\alias{visualize_area}
\title{visualize integration area}
\usage{
visualize_area(bounds, constrains)
}
\arguments{
\item{bounds:necessary, }{the bounds of x, y, z, as a vector of 2, 4 or 6 dimensions}

\item{constrains:}{optional, inequalities that x, y, z must satisfy, input as character, must contains x, y, or z, default: 1}
}
\description{
Visualize 2D/3D integration area with random experienments
}
\examples{
visualize_area(c(0,1, 1,2, -1,3), c("exp(x)<y", "sin(z)*y>=0"))

}
