% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmi.R
\name{mcmi}
\alias{mcmi}
\title{MCMI}
\usage{
mcmi(bounds, integrand, constrains)
}
\arguments{
\item{Bounds:}{the bounds of x, y, z, as a vector of 2, 4 or 6 dimensions}

\item{Integrand:}{the integrand}

\item{Constrains:}{inequalities that x, y, z must satisfy}
}
\description{
Calculate multivariable integration using Monte Carlo method
}
\examples{
mcmi(c(2,3, 3,4), x^2+y^2, x<y)

}
